Move db::item_type_if_contained to ActionTesting.hpp
It is now in a detail namespace as it is only used in the
implementation of action testing.
